+++ alicloud/resource_alicloud_db_database_test.go
@@ -56,10 +56,6 @@ func testAccCheckDBDatabaseExists(n string, d *rds.Database) resource.TestCheckF
 			return err
 		}
 
-		if db == nil {
-			return fmt.Errorf("DB is not found in the instance %s.", parts[0])
-		}
-
 		*d = *db
 		return nil
 	}
@@ -75,19 +71,14 @@ func testAccCheckDBDatabaseDestroy(s *terraform.State) error {
 
 		parts := strings.Split(rs.Primary.ID, COLON_SEPARATED)
 
-		db, err := client.DescribeDatabaseByName(parts[0], parts[1])
-
-		// Verify the error is what we want
-		if err != nil {
-			if NotFoundDBInstance(err) || IsExceptedError(err, InvalidDBNameNotFound) {
+		if _, err := client.DescribeDatabaseByName(parts[0], parts[1]); err != nil {
+			if NotFoundError(err) {
 				continue
 			}
 			return err
 		}
 
-		if db != nil {
-			return fmt.Errorf("Error database %s is still existing.", parts[1])
-		}
+		return fmt.Errorf("Error database %s is still existing.", parts[1])
 	}
 
 	return nil