+++ CHANGELOG.md
@@ -9,6 +9,10 @@ IMPROVEMENTS:
 - Base64 encode ESS scaling config user_data ([#315](https://github.com/terraform-providers/terraform-provider-alicloud/pull/315))
 - Retrieve the account_id automatically if needed ([#314](https://github.com/terraform-providers/terraform-provider-alicloud/pull/314))
 
+BUG FIXES:
+
+- Fix DB database not found error ([#316](https://github.com/terraform-providers/terraform-provider-alicloud/pull/316))
+
 ## 1.14.0 (August 31, 2018)
 
 FEATURES: